Got SDL2 build compiling again.. Not sure if it works yet, crashes KDE on my VM

+++ Source/Core/JpegTask.cpp
@@ -140,12 +140,12 @@ void jpeg_decode_PS(OSTask *task)
         return;
     }
     #endif
-    u32       address          {rdram_read_u32((u32)task->t.data_ptr)};
-    const u32 macroblock_count {rdram_read_u32((u32)task->t.data_ptr + 4)};
-    const u32 mode             {rdram_read_u32((u32)task->t.data_ptr + 8)};
-    const u32 qtableY_ptr      {rdram_read_u32((u32)task->t.data_ptr + 12)};
-    const u32 qtableU_ptr      {rdram_read_u32((u32)task->t.data_ptr + 16)};
-    const u32 qtableV_ptr      {rdram_read_u32((u32)task->t.data_ptr + 20)};
+    u32       address          {rdram_read_u32((uintptr_t)task->t.data_ptr)};
+    const u32 macroblock_count {rdram_read_u32((uintptr_t)task->t.data_ptr + 4)};
+    const u32 mode             {rdram_read_u32((uintptr_t)task->t.data_ptr + 8)};
+    const u32 qtableY_ptr      {rdram_read_u32((uintptr_t)task->t.data_ptr + 12)};
+    const u32 qtableU_ptr      {rdram_read_u32((uintptr_t)task->t.data_ptr + 16)};
+    const u32 qtableV_ptr      {rdram_read_u32((uintptr_t)task->t.data_ptr + 20)};
 
     #ifdef DAEDALUS_DEBUG_CONSOLE
     if (mode != 0 && mode != 2)
@@ -203,7 +203,7 @@ void jpeg_decode_OB(OSTask *task)
 
     s32 y_dc {}, u_dc {}, v_dc {};
 
-	u32  address  {(u32)task->t.data_ptr};
+	u32  address  {(uintptr_t)task->t.data_ptr};
 	const u32 macroblock_count {task->t.data_size};
 	const u32  qscale   {task->t.yield_data_size};
