+++ Source/Core/RSP_HLE.cpp
@@ -205,7 +205,7 @@ void jpeg_decode_OB(OSTask *task);
 
 	// most ucode_boot procedure copy 0xf80 bytes of ucode whatever the ucode_size is.
 	// For practical purpose we use a ucode_size = min(0xf80, task->ucode_size)
-	u32 sum {sum_bytes(g_pu8RamBase + (u32)task->t.ucode , Min<u32>(task->t.ucode_size, 0xf80) >> 1)};
+	u32 sum {sum_bytes(g_pu8RamBase + (uintptr_t)task->t.ucode , Min<u32>(task->t.ucode_size, 0xf80) >> 1)};
 
 	//DBGConsole_Msg(0, "JPEG Task: Sum=0x%08x", sum);
 	switch(sum)
