+++ Source/DynaRec/AssemblyUtils.h
@@ -38,7 +38,7 @@ public:
 	bool			IsSet() const				{ return mpLocation != nullptr; }
 	const void *	GetTarget() const			{ return mpLocation; }
 	const u8 *		GetTargetU8P() const		{ return reinterpret_cast< const u8 * >( mpLocation ); }
-	u32				GetTargetU32() const		{ return reinterpret_cast< u32 >( mpLocation ); }
+	u32				GetTargetU32() const		{ return reinterpret_cast< uintptr_t >( mpLocation ); }
