+++ Source/HLEGraphics/RDPStateManager.h
@@ -91,7 +91,7 @@ extern CRDPStateManager		gRDPStateManager;
 extern RDP_OtherMode		gRDPOtherMode;
 
 extern u32* gTlutLoadAddresses[ 4096 >> 6 ];
-#define TLUT_BASE ((u32)(gTlutLoadAddresses[0]))
+#define TLUT_BASE ((uintptr_t)(gTlutLoadAddresses[0]))
 
 
 #endif // HLEGRAPHICS_RDPSTATEMANAGER_H_
